# backend/docs/openapi.yaml
openapi: 3.1.0
info:
  title: API CineVault
  description: Documentación oficial de la API de CineVault (red social de cine).
  version: 1.0.0
  contact:
    name: Equipo CineVault
servers:
  - url: https://api.cinevault.art/api
    description: Producción
  - url: http://localhost:4000/api
    description: Desarrollo local
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    cookieAuth:
      type: apiKey
      in: cookie
      name: refresh_token
  schemas:
...
    Favorito:
      type: object
      properties:
        movie_id: { type: integer }
        rank_position: { type: integer, nullable: true }
...
    ErrorResponse:
      type: object
      properties:
        error: { type: string }
        message: { type: string }
        code: { type: string }
...
  responses:
...
    NotFound:
      description: Recurso no encontrado
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
paths:
...
  /favorites:
    get:
      tags: [Favoritos]
      summary: Mis favoritos
      security: [{ bearerAuth: [] }]
      responses:
        '200': { description: Lista, content: { application/json: { schema: { type: array, items: { $ref: '#/components/schemas/Favorito' } } } } }
  /favorites/user/{userId}:
    get:
      tags: [Favoritos]
      summary: Favoritos de un usuario
      parameters:
        - in: path
          name: userId
          required: true
          schema: { type: integer }
      responses:
        '200': { description: Lista }
  /favorites/{movieId}:
    post:
      tags: [Favoritos]
      summary: Añadir favorito
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: movieId
          required: true
          schema: { type: integer }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                rank_position: { type: integer }
      responses:
        '201': { description: Añadido }
    delete:
      tags: [Favoritos]
      summary: Eliminar favorito
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: movieId
          required: true
          schema: { type: integer }
      responses:
        '200': { description: Eliminado }
...
  /reviews/movie/{movieId}:
    get:
      tags: [Reseñas]
      summary: Reseñas por película
      parameters:
        - in: path
          name: movieId
          required: true
          schema: { type: integer }
      responses:
        '200': { description: Lista }
...
  /search/multi:
    get:
      tags: [Búsqueda]
      summary: Búsqueda múltiple
      parameters:
        - in: query
          name: q
          required: true
          schema: { type: string }
      responses:
        '200': { description: Resultados }
  /search/movie:
    get:
      tags: [Búsqueda]
      summary: Búsqueda de películas
      parameters:
        - in: query
          name: q
          required: true
          schema: { type: string }
      responses:
        '200': { description: Resultados }
  /search/person:
    get:
      tags: [Búsqueda]
      summary: Búsqueda de personas
      parameters:
        - in: query
          name: q
          required: true
          schema: { type: string }
      responses:
        '200': { description: Resultados }
  /search/tv:
    get:
      tags: [Búsqueda]
      summary: Búsqueda de series
      parameters:
        - in: query
          name: q
          required: true
          schema: { type: string }
      responses:
        '200': { description: Resultados }
...
  /movies/{idOrSlug}:
    get:
      tags: [Películas]
      summary: Detalle por ID o slug
      parameters:
        - in: path
          name: idOrSlug
          required: true
          schema: { type: string }
      responses:
        '200': { description: Detalle }
        '404': { $ref: '#/components/responses/NotFound' }
...